third_party/blink/web_tests/fast/multicol/columns-shorthand-parsing-2-expected.txt

Test the behavior when 'auto' is part of the 'columns' property value. See http://www.w3.org/TR/css3-multicol/#columns

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS element.style.WebkitColumnWidth is "auto"
PASS element.style.WebkitColumnCount is "3"
PASS element.style.WebkitColumns is "3"
PASS element.style.WebkitColumnWidth is "10em"
PASS element.style.WebkitColumnCount is "auto"
PASS element.style.WebkitColumns is "10em"
PASS element.style.WebkitColumnWidth is "auto"
PASS element.style.WebkitColumnCount is "3"
PASS element.style.WebkitColumns is "3"
PASS element.style.WebkitColumnWidth is "10em"
PASS element.style.WebkitColumnCount is "auto"
PASS element.style.WebkitColumns is "10em"
PASS element.style.WebkitColumnWidth is "7em"
PASS element.style.WebkitColumnCount is "7"
PASS element.style.WebkitColumns is "7em 7"
PASS element.style.WebkitColumnWidth is "7em"
PASS element.style.WebkitColumnCount is "7"
PASS element.style.WebkitColumns is "7em 7"
PASS element.style.WebkitColumnWidth is "auto"
PASS element.style.WebkitColumnCount is "auto"
PASS element.style.WebkitColumns is "auto"
PASS element.style.WebkitColumnWidth is "auto"
PASS element.style.WebkitColumnCount is "auto"
PASS element.style.WebkitColumns is "auto"
PASS element.style.WebkitColumnWidth is "initial"
PASS element.style.WebkitColumnCount is "initial"
PASS element.style.WebkitColumns is "initial"
PASS element.style.WebkitColumnWidth is "auto"
PASS element.style.WebkitColumnCount is "auto"
PASS element.style.WebkitColumns is "auto"
PASS element.style.WebkitColumnWidth is "inherit"
PASS element.style.WebkitColumnCount is "inherit"
PASS element.style.WebkitColumns is "inherit"
PASS element.style.WebkitColumnWidth is "auto"
PASS element.style.WebkitColumnCount is "auto"
PASS element.style.WebkitColumns is "auto"
PASS element.style.WebkitColumnWidth is "auto"
PASS element.style.WebkitColumnCount is "7"
PASS element.style.WebkitColumns is "7"
PASS element.style.WebkitColumnWidth is "7em"
PASS element.style.WebkitColumnCount is "auto"
PASS element.style.WebkitColumns is "7em"
PASS successfullyParsed is true

TEST COMPLETE
